+++ pueblo/ngr/core.py
@@ -19,6 +19,7 @@ from pathlib import Path
 from pueblo.ngr.model import ItemType
 from pueblo.ngr.runner import (
     DotNetRunner,
+    GolangRunner,
     JavaRunner,
     JavaScriptRunner,
     JuliaRunner,
@@ -47,6 +48,7 @@ class NextGenerationRunner:
         self.runners = {
             # Language runners, listed alphabetically.
             ItemType.DOTNET: DotNetRunner,
+            ItemType.GOLANG: GolangRunner,
             ItemType.JAVA: JavaRunner,
             ItemType.JAVASCRIPT: JavaScriptRunner,
             ItemType.JULIA: JuliaRunner,